refactor:whitelist extension to use refactored components
